pyproject.toml
[tool.poetry]
name = "inverted-pendulum-dqn"
version = "0.1.0"
description = "Deep Q-Network for Inverted Pendulum Control with TensorFlow (CPU)"
authors = []
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
tensorflow = ">=2.13.0"
numpy = ">=1.24.0,<2"
matplotlib = ">=3.7.0"
pyyaml = ">=6.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
